feat: 로그인 사용자 요금 계산 정책 테스트 추가 및 요금 타입 변경

+++ src/main/java/wooteco/subway/path/domain/SubwayPath.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class SubwayPath {
     private static final int BASIC_FARE = 1250;
     private static final int OVER_FARE = 2050;
-    private static final int FARE_PER_KM = 100;
+    private static final double FARE_PER_KM = 100;
     private static final int FIRST_OVER_FARE_DISTANCE = 10;
     private static final int SECOND_OVER_FARE_DISTANCE = 50;
 
@@ -19,10 +19,10 @@ public class SubwayPath {
         this.stations = stations;
     }
 
-    public int calculateFare() {
+    public double calculateFare() {
         int distance = calculateDistance();
         long extraFare = findExtraFare();
-        int baseFare = BASIC_FARE + (int) extraFare;
+        double baseFare = BASIC_FARE + extraFare;
 
         if (distance <= FIRST_OVER_FARE_DISTANCE) {
             return baseFare;
@@ -32,7 +32,7 @@ public class SubwayPath {
             return calculateOverFare(baseFare, distance - FIRST_OVER_FARE_DISTANCE, 5);
         }
 
-        baseFare = OVER_FARE + (int) extraFare;
+        baseFare = OVER_FARE + extraFare;
         return calculateOverFare(baseFare, distance - SECOND_OVER_FARE_DISTANCE, 8);
     }
 
@@ -44,8 +44,8 @@ public class SubwayPath {
                 .orElse(0);
     }
 
-    private int calculateOverFare(int baseFare, int distance, int perDistance) {
-        return baseFare + (int) (Math.ceil(distance / perDistance) * FARE_PER_KM);
+    private double calculateOverFare(double baseFare, int distance, int perDistance) {
+        return baseFare + (Math.ceil(distance / perDistance) * FARE_PER_KM);
     }
 
     public List<SectionEdge> getSectionEdges() {
