+++ src/main/java/wooteco/subway/path/dto/PathResponse.java
@@ -7,12 +7,12 @@ import java.util.List;
 public class PathResponse {
     private List<StationResponse> stations;
     private int distance;
-    private int fare;
+    private double fare;
 
     public PathResponse() {
     }
 
-    public PathResponse(List<StationResponse> stations, int distance, int fare) {
+    public PathResponse(List<StationResponse> stations, int distance, double fare) {
         this.stations = stations;
         this.distance = distance;
         this.fare = fare;
@@ -26,7 +26,7 @@ public class PathResponse {
         return distance;
     }
 
-    public int getFare() {
+    public double getFare() {
         return fare;
     }
 }
